+++ Manager/Properties/AssemblyInfo.cs
@@ -51,6 +51,6 @@ using System.Windows;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion( "3.0.2021.0111" )]
-[assembly: AssemblyFileVersion( "3.0.2021.0111" )]
+[assembly: AssemblyVersion( "3.0.2021.0120" )]
+[assembly: AssemblyFileVersion( "3.0.2021.0120" )]
 [assembly: NeutralResourcesLanguage( "en" ) ]